ConnectionServiceException and UserAlreadyReservedException moved to the device-registry-api module
